+++ src/components/pages/ComicsPage.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import ComicsList from '../comicsList/ComicsList'
+
+function ComicsPage() {
+    return <ComicsList />
+}
+
+export default ComicsPage
