+++ src/components/pages/index.ts
@@ -1,0 +1,4 @@
+import MainPage from './MainPage'
+import ComicsPage from './ComicsPage'
+
+export { MainPage, ComicsPage }
